refactor: reorganize namespaces and add new folder
- Added a new folder `<Folder Include="Integration\Attributes\" />` in `EasyAuth.csproj`.
- Updated namespace imports in `EasyAuthExtension.cs` to include `EasyAuth.Integration.Middleware`.
- Changed namespace from `EasyAuth.Integration` to `EasyAuth.Integration.Middleware` in `PermissionHandler.cs`.
- Updated namespace in `PermissionRequirement.cs` and improved code readability by removing unnecessary `this.` in the assignment of `Permission`.
