+++ EasyAuth.csproj
@@ -10,4 +10,8 @@
     <PackageReference Include="Microsoft.AspNetCore.Authorization" Version="9.0.4" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Integration\Attributes\" />
+  </ItemGroup>
+
 </Project>
